[s-1] Security token service: 1. Created own Exception for incorrect login parameters 2. Created entity Session with user token 3. Created SecurityService and getUserByLogin method for authentication 4. Created tests for this functions 5. Change users parameters from data.sql 6. Change type phoneNumber from int to Long 7. Created fragment head.html

+++ src/main/java/com/photostudio/web/servlet/user/UserServlet.java
@@ -37,7 +37,7 @@ public class UserServlet extends HttpServlet {
         User newUser = new User();
 
         newUser.setEmail(email);
-        newUser.setPhoneNumber(Integer.parseInt(phoneNumber));
+        newUser.setPhoneNumber(Long.parseLong(phoneNumber));
         newUser.setGender(Gender.getByGender(gender));
         newUser.setFirstName(firstName);
         newUser.setLastName(lastName);
